Fixes error in checkout
During checkout, Shipment#manifest_unstock calls StockLocation#unstock to remove sold items from stock. Checkout was breaking for reserved items because it wasn’t passing in a user to find the right ReservedStockItem for.
I’d also required an original_stock_location argument to find a particular ReservedStockItem. In this case, though, it doesn’t matter so much which stock item is unstocked, provided that we get the right user and variant. So I’ve relaxed that condition. If StockLocation#stock_item receives an original_stock_location_id argument it’ll use it, otherwise it’ll return the first relevant item it finds.

- adds shipment_decorator
- removes error for missing original_stock_location in StockLocation#stock_item
- various code style improvements
- tiny version bump