Adds validation forbidding regular StockItems in reserved stock locations
Before this validation, it was possible to create a regular stock item on a reserved stock location. This caused some problems because other parts of the code (e.g. show_user_for_reserved_stock.html.erb.deface) assume that the stock items for a reserved stock location will always be ReservedStockItems.